services/tenant-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.0
redis==5.0.0
aioredis==2.0.1
python-json-logger==2.0.7
pyjwt==2.9.0
email-validator==2.2.0
